perf: 优化应用树，统一 dropdown
